Backport from trunk rev6852: Feature: add an option to clean_dump_archive or not (default is True)
git-svn-id: file:///home/code/svn/ovd/branches/3.0@7749 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ OvdServer/ovd/Role/ApplicationServer/Role.py
@@ -94,7 +94,8 @@ class Role(AbstractRole):
 			Logger.error("Unable to purge group")
 			return False
 		
-		self.purgeArchives()
+		if Config.clean_dump_archive:
+			self.purgeArchives()
 		
 		if Config.multithread:
 			cpuInfos = Platform.System.getCPUInfos()
